# feat: Implement concurrent directory operations for List command