--- a/internal/ticket/manager.go
+++ b/internal/ticket/manager.go
@@ -7,8 +7,13 @@ import (
 	"io"
 	"os"
 	"path/filepath"
+	"runtime"
 	"sort"
 	"strings"
+	"sync"
+
+	"golang.org/x/sync/errgroup"
+	"golang.org/x/sync/semaphore"
 
 	"github.com/yshrsmz/ticketflow/internal/config"
 	ticketerrors "github.com/yshrsmz/ticketflow/internal/errors"
@@ -124,6 +129,36 @@ func (m *Manager) List(ctx context.Context, statusFilter StatusFilter) ([]Ticket
 		return nil, fmt.Errorf("invalid status filter: %s", statusFilter)
 	}
 
+	// Count total files first to better estimate capacity
+	totalFiles := 0
+	for _, dir := range dirs {
+		entries, err := os.ReadDir(dir)
+		if err != nil {
+			if !os.IsNotExist(err) {
+				// Log error but continue with other directories
+				continue
+			}
+			continue
+		}
+		for _, entry := range entries {
+			if !entry.IsDir() && strings.HasSuffix(entry.Name(), ".md") {
+				totalFiles++
+			}
+		}
+	}
+
+	// Use concurrent loading if we have enough files to benefit from it
+	// Threshold is set to 10 files based on typical overhead of goroutines
+	if totalFiles >= 10 {
+		return m.listConcurrent(ctx, dirs)
+	}
+
+	// Fall back to sequential for small numbers of tickets
+	return m.listSequential(ctx, dirs)
+}
+
+// listSequential lists tickets sequentially (original implementation)
+func (m *Manager) listSequential(ctx context.Context, dirs []string) ([]Ticket, error) {
 	// Pre-allocate tickets slice with reasonable capacity based on typical usage
 	// This avoids multiple reallocations during append operations without double-reading directories
 	tickets := make([]Ticket, 0, initialTicketCapacity)
@@ -154,6 +189,97 @@ func (m *Manager) List(ctx context.Context, statusFilter StatusFilter) ([]Ticket
 
 			tickets = append(tickets, *ticket)
 		}
+	}
+
+	// Sort by priority first, then by creation time (newest first)
+	sort.Slice(tickets, func(i, j int) bool {
+		if tickets[i].Priority != tickets[j].Priority {
+			return tickets[i].Priority < tickets[j].Priority
+		}
+		return tickets[i].CreatedAt.After(tickets[j].CreatedAt.Time)
+	})
+
+	return tickets, nil
+}
+
+// listConcurrent lists tickets using concurrent file operations
+func (m *Manager) listConcurrent(ctx context.Context, dirs []string) ([]Ticket, error) {
+	// Collect all ticket files to process
+	var ticketPaths []string
+	for _, dir := range dirs {
+		entries, err := os.ReadDir(dir)
+		if err != nil {
+			if os.IsNotExist(err) {
+				continue
+			}
+			return nil, fmt.Errorf("failed to read directory %s: %w", dir, err)
+		}
+
+		for _, entry := range entries {
+			if !entry.IsDir() && strings.HasSuffix(entry.Name(), ".md") {
+				ticketPaths = append(ticketPaths, filepath.Join(dir, entry.Name()))
+			}
+		}
+	}
+
+	if len(ticketPaths) == 0 {
+		return []Ticket{}, nil
+	}
+
+	// Determine optimal number of workers
+	numWorkers := runtime.NumCPU()
+	if numWorkers > len(ticketPaths) {
+		numWorkers = len(ticketPaths)
+	}
+	// Cap at 8 workers to avoid excessive file handles
+	if numWorkers > 8 {
+		numWorkers = 8
+	}
+
+	// Create semaphore to limit concurrent file operations
+	sem := semaphore.NewWeighted(int64(numWorkers))
+
+	// Pre-allocate result slice with exact capacity
+	tickets := make([]Ticket, 0, len(ticketPaths))
+	var mu sync.Mutex
+
+	// Use errgroup for structured concurrency
+	g, ctx := errgroup.WithContext(ctx)
+
+	for _, path := range ticketPaths {
+		path := path // Capture for goroutine
+		
+		g.Go(func() error {
+			// Acquire semaphore
+			if err := sem.Acquire(ctx, 1); err != nil {
+				return fmt.Errorf("failed to acquire semaphore: %w", err)
+			}
+			defer sem.Release(1)
+
+			// Check context before loading
+			if err := ctx.Err(); err != nil {
+				return fmt.Errorf("operation cancelled: %w", err)
+			}
+
+			// Load ticket
+			ticket, err := m.loadTicket(ctx, path)
+			if err != nil {
+				// Skip invalid tickets - don't fail the entire operation
+				return nil
+			}
+
+			// Add to results with mutex protection
+			mu.Lock()
+			tickets = append(tickets, *ticket)
+			mu.Unlock()
+
+			return nil
+		})
+	}
+
+	// Wait for all goroutines to complete
+	if err := g.Wait(); err != nil {
+		return nil, err
 	}
 
 	// Sort by priority first, then by creation time (newest first)