--- a/internal/ticket/manager_benchmark_test.go
+++ b/internal/ticket/manager_benchmark_test.go
@@ -356,6 +356,92 @@ func BenchmarkManagerListConcurrent(b *testing.B) {
 	})
 }
 
+// BenchmarkManagerListSequentialVsConcurrent compares sequential vs concurrent list implementations
+func BenchmarkManagerListSequentialVsConcurrent(b *testing.B) {
+	scenarios := []struct {
+		name        string
+		ticketCount int
+	}{
+		{"10-tickets", 10},
+		{"50-tickets", 50},
+		{"100-tickets", 100},
+		{"200-tickets", 200},
+		{"500-tickets", 500},
+	}
+
+	for _, scenario := range scenarios {
+		b.Run(scenario.name, func(b *testing.B) {
+			// Setup
+			tmpDir := b.TempDir()
+			cfg := config.Default()
+			cfg.Tickets.Dir = "tickets"
+			manager := NewManager(cfg, tmpDir)
+			ctx := context.Background()
+
+			// Create ticket directories
+			todoDir := filepath.Join(tmpDir, cfg.Tickets.Dir, "todo")
+			doingDir := filepath.Join(tmpDir, cfg.Tickets.Dir, "doing")
+			doneDir := filepath.Join(tmpDir, cfg.Tickets.Dir, "done")
+			for _, dir := range []string{todoDir, doingDir, doneDir} {
+				if err := os.MkdirAll(dir, 0755); err != nil {
+					b.Fatal(err)
+				}
+			}
+
+			// Create tickets distributed across statuses
+			for i := 0; i < scenario.ticketCount; i++ {
+				slug := fmt.Sprintf("benchmark-ticket-%d", i)
+				ticket, err := manager.Create(ctx, slug)
+				if err != nil {
+					b.Fatal(err)
+				}
+
+				// Move some tickets to different statuses
+				switch i % 3 {
+				case 1: // Move to doing
+					oldPath := filepath.Join(todoDir, ticket.ID+".md")
+					newPath := filepath.Join(doingDir, ticket.ID+".md")
+					if err := os.Rename(oldPath, newPath); err != nil {
+						b.Fatal(err)
+					}
+				case 2: // Move to done
+					oldPath := filepath.Join(todoDir, ticket.ID+".md")
+					newPath := filepath.Join(doneDir, ticket.ID+".md")
+					if err := os.Rename(oldPath, newPath); err != nil {
+						b.Fatal(err)
+					}
+				}
+			}
+
+			dirs := []string{todoDir, doingDir, doneDir}
+
+			// Benchmark sequential implementation
+			b.Run("sequential", func(b *testing.B) {
+				b.ReportAllocs()
+				b.ResetTimer()
+				for i := 0; i < b.N; i++ {
+					_, err := manager.listSequential(ctx, dirs)
+					if err != nil {
+						b.Fatal(err)
+					}
+				}
+			})
+
+			// Benchmark concurrent implementation
+			b.Run("concurrent", func(b *testing.B) {
+				b.ReportAllocs()
+				b.ResetTimer()
+				for i := 0; i < b.N; i++ {
+					_, err := manager.listConcurrent(ctx, dirs)
+					if err != nil {
+						b.Fatal(err)
+					}
+				}
+			})
+		})
+	}
+}
+
 // BenchmarkManagerCurrentTicket benchmarks getting and setting current ticket
 func BenchmarkManagerCurrentTicket(b *testing.B) {
 	tmpDir := b.TempDir()