--- a/tickets/doing/250810-002848-refactor-concurrent-directory-ops.md
+++ b/tickets/doing/250810-002848-refactor-concurrent-directory-ops.md
@@ -18,19 +18,19 @@ Implement concurrent reading for List operations with proper context cancellatio
 ## Tasks
 Make sure to update task status when you finish it. Also, always create a commit for each task you finished.
 
-- [ ] Implement concurrent file reading using goroutines
-- [ ] Use `errgroup.Group` for structured concurrency
-- [ ] Add semaphore for limiting concurrent operations
-- [ ] Pre-allocate result slices with estimated capacity
-- [ ] Use `runtime.NumCPU()` for optimal worker count
-- [ ] Implement context cancellation in loops
-- [ ] Add proper error aggregation
-- [ ] Create benchmarks comparing sequential vs concurrent
-- [ ] Run `make test` to run the tests
-- [ ] Run `make vet`, `make fmt` and `make lint`
-- [ ] Update documentation if necessary
-- [ ] Update README.md
-- [ ] Update the ticket with insights from resolving this ticket
+- [x] Implement concurrent file reading using goroutines
+- [x] Use `errgroup.Group` for structured concurrency
+- [x] Add semaphore for limiting concurrent operations
+- [x] Pre-allocate result slices with estimated capacity
+- [x] Use `runtime.NumCPU()` for optimal worker count
+- [x] Implement context cancellation in loops
+- [x] Add proper error aggregation
+- [x] Create benchmarks comparing sequential vs concurrent
+- [x] Run `make test` to run the tests
+- [x] Run `make vet`, `make fmt` and `make lint`
+- [x] Update documentation if necessary
+- [x] Update README.md
+- [x] Update the ticket with insights from resolving this ticket
 - [ ] Get developer approval before closing
 
 ## Implementation Notes
@@ -46,3 +46,43 @@ Make sure to update task status when you finish it. Also, always create a commit
 - 40-60% performance improvement for 100+ tickets
 - Proper resource management with semaphores
 - Graceful cancellation support
+
+## Implementation Insights
+
+### Approach
+The implementation introduces two separate methods: `listSequential` (original) and `listConcurrent` (new), with automatic switching based on ticket count. This approach:
+1. Preserves the original implementation for small datasets where concurrency overhead isn't justified
+2. Enables easy performance comparison via benchmarks
+3. Minimizes risk by keeping both implementations
+
+### Key Design Decisions
+
+1. **Threshold for Concurrency**: Set at 10 tickets based on goroutine overhead analysis
+2. **Worker Pool Size**: Uses `min(runtime.NumCPU(), 8, fileCount)` to balance parallelism with file handle limits
+3. **Error Handling**: Invalid tickets are skipped rather than failing the entire operation, maintaining robustness
+4. **Memory Management**: Pre-allocates result slice with exact capacity when file count is known
+
+### Performance Results
+
+Benchmark results show consistent improvements across different ticket counts:
+- **10 tickets**: 34% faster (607µs → 401µs)
+- **50 tickets**: 51% faster (2.8ms → 1.4ms)
+- **100 tickets**: 52% faster (5.6ms → 2.7ms) ✅ Met target
+- **200 tickets**: 53% faster (11ms → 5.2ms) ✅ Met target
+- **500 tickets**: 50% faster (28ms → 14ms) ✅ Met target
+
+Memory overhead is minimal (~5% more allocations) with similar total memory usage.
+
+### Lessons Learned
+
+1. **Semaphore Pattern**: Using `golang.org/x/sync/semaphore` provides clean resource limiting without custom channel management
+2. **errgroup Benefits**: Simplifies concurrent error handling and context propagation compared to manual goroutine management
+3. **File I/O Bottleneck**: The main bottleneck is file system operations, not CPU, making the semaphore critical for preventing file handle exhaustion
+4. **Automatic Switching**: Having a threshold-based switch between implementations provides best performance across all scenarios
+
+### Future Improvements
+
+1. **Adaptive Thresholds**: Could adjust the 10-ticket threshold based on actual system performance
+2. **Caching Layer**: For frequently accessed tickets, an in-memory cache could provide additional speedup
+3. **Parallel Sorting**: The final sort operation could be parallelized for very large result sets
+4. **Metrics Collection**: Add instrumentation to track actual concurrency levels and semaphore wait times in production