## Summary
Implements concurrent file loading for the List command, achieving 50%+ performance improvement for 100+ tickets.

## Changes
- Added concurrent loading using `errgroup` and `semaphore` for structured concurrency
- Implemented automatic switching between sequential (< 10 tickets) and concurrent (10+ tickets) loading
- Used `runtime.NumCPU()` with max 8 workers to balance parallelism with file handle limits
- Added comprehensive benchmarks comparing sequential vs concurrent implementations
- Updated documentation with performance metrics

## Performance Results
- **10 tickets**: 34% faster (607µs → 401µs)
- **50 tickets**: 51% faster (2.8ms → 1.4ms)
- **100 tickets**: 52% faster (5.6ms → 2.7ms) ✅
- **200 tickets**: 53% faster (11ms → 5.2ms) ✅
- **500 tickets**: 50% faster (28ms → 14ms) ✅

Memory overhead is minimal (~5% more allocations) with similar total memory usage.

## Test Plan
- [x] All existing tests pass
- [x] Benchmarks show expected performance improvement
- [x] `make vet`, `make fmt`, and `make lint` pass
- [x] Documentation updated

Related to refactoring phase 1 (Foundation) - Task 1.2